requirements.txt
nicegui>=2.0,<3.0
uvicorn[standard]>=0.30
psycopg[binary]>=3.1
python-dotenv>=1.0
bcrypt>=4.0,<5
